+++ src/pages/legacy/overallSchedule/utils/getFormattedAvailableDateTimes.ts
