+++ src/pages/selectSchedule/SelectSchedule.tsx
@@ -11,26 +11,12 @@ import Description from './components/Description';
 import SelectScheduleTable from './components/SelectScheduleTable';
 import { ScheduleStepContext } from './context';
 import { ScheduleStepType } from './types';
+import { TITLES } from './utils';
 
 function SelectSchedule() {
   const [scheduleStep, setScheduleStep] = useState<ScheduleStepType>('selectTimeSlot');
   const { meetingId } = useParams();
   const { data, isLoading } = useGetTimetable(meetingId);
-  interface TitlesType {
-    [key: string]: {
-      main: string;
-      sub?: string;
-    };
-  }
-  const titles: TitlesType = {
-    selectTimeSlot: {
-      main: '가능한 시간대를 등록해주세요',
-      sub: '시작시간과 종료시간을 터치하여 블럭을 생성해주세요',
-    },
-    selectPriority: {
-      main: '우선순위를 입력해주세요',
-    },
-  };
 
   return (
     <ScheduleStepContext.Provider value={{ scheduleStep, setScheduleStep }}>
@@ -46,8 +32,8 @@ function SelectSchedule() {
               />
             )}
             <TitleComponents
-              main={titles[scheduleStep].main}
-              sub={titles[scheduleStep].sub}
+              main={TITLES[scheduleStep].main}
+              sub={TITLES[scheduleStep].sub}
               padding={scheduleStep === 'selectTimeSlot' ? `0 0 2.6rem` : `4.4rem 0 3.2rem 0`}
             />
             <SelectScheduleTable
