+++ src/pages/selectSchedule/types.ts
@@ -5,3 +5,10 @@ import { ColumnStructure } from 'components/timetableComponents/types';
 export type ScheduleStepType = 'selectTimeSlot' | 'selectPriority';
 export type StepSlotsType = { [key in ScheduleStepType]: (props: ColumnStructure) => ReactNode };
 export type StepbottomItemsType = { [key in ScheduleStepType]: ReactNode };
+
+export interface TitlesType {
+  [key: string]: {
+    main: string;
+    sub?: string;
+  };
+}
